coisarada

+++ Codigos/Matemática/Primos/naive_is_prime.cpp
@@ -1,5 +1,5 @@
 bool is_prime(int n) {
-    for (long long d = 2; d * d <= n; d++) {
+    for (int d = 2; d * d <= n; d++) {
         if (n % d == 0) {
             return false;
         }
